+++ src/main/java/radioactivity/entities/Enemy.java
@@ -2,7 +2,7 @@ package radioactivity.entities;
 
 public class Enemy extends BaseCharacter {
 
-    public Enemy(String name, int life, int damage, int velocity) {
+    public Enemy(String name, int life, int damage, double velocity) {
         super(name, life, damage, velocity);
     }
